requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-generativeai>=0.8.3
python-dotenv==1.0.1
requests==2.31.0
